refactor: extract config logic to dedicated module
- Create new src/config/index.js with all config-related functions
- Move resolveConfigPath, validateConfigStructure, readConfig, pickServer from mcp.js
- Move getClaudeConfigPath from utils.js to config module
- Add backward compatibility re-export in utils.js
- Improve separation of concerns and maintainability
- All existing functionality preserved
