+++ src/utils.js
@@ -1,12 +1,13 @@
 import traverse from '@json-schema-tools/traverse'
 import { isEmpty, set as setPath, truncate } from 'lodash-es'
 import { Console } from 'node:console'
-import { homedir } from 'os'
-import { join } from 'path'
 import prompts from 'prompts'
 import { parse } from 'uri-template'
 import yoctoSpinner from 'yocto-spinner'
 import colors from 'yoctocolors'
+
+// Re-export for backward compatibility
+export { getClaudeConfigPath } from './config/index.js'
 
 export const logger = new Console({ stdout: process.stderr, stderr: process.stderr })
 
@@ -18,14 +19,6 @@ export function createSpinner(text) {
   return yoctoSpinner({ text, stream: process.stderr }).start()
 }
 
-export function getClaudeConfigPath() {
-  if (process.platform === 'win32') {
-    return join(homedir(), 'AppData', 'Roaming', 'Claude', 'claude_desktop_config.json')
-  }
-  if (process.platform === 'darwin') {
-    return join(homedir(), 'Library', 'Application Support', 'Claude', 'claude_desktop_config.json')
-  }
-}
 
 export async function readPromptArgumentInputs(args) {
   if (!args || args.length === 0) {
